Refactor messages
The old Message component used a css animation as a close animation. It wasn't flexible and didn't animate the enter of the message, as well as hiding Message related variables in index.css. Now it uses a transition based on a time stored in a state in Message. It allow multiple steps in the lifetime of the component. The component is also removed at the end of the animation, which wasn't the case before.